# Impl Extend for LiteMap and avoid quadratic behavior in from_iter and deserialize