--- a/utils/litemap/src/lib.rs
+++ b/utils/litemap/src/lib.rs
@@ -7,13 +7,27 @@
 //! `litemap` is a crate providing [`LiteMap`], a highly simplistic "flat" key-value map
 //! based off of a single sorted vector.
 //!
-//! The goal of this crate is to provide a map that is good enough for small
+//! The main goal of this crate is to provide a map that is good enough for small
 //! sizes, and does not carry the binary size impact of [`HashMap`](std::collections::HashMap)
 //! or [`BTreeMap`](alloc::collections::BTreeMap).
 //!
 //! If binary size is not a concern, [`std::collections::BTreeMap`] may be a better choice
 //! for your use case. It behaves very similarly to [`LiteMap`] for less than 12 elements,
 //! and upgrades itself gracefully for larger inputs.
+//!
+//! ## Performance characteristics
+//!
+//! [`LiteMap`] is a data structure with similar characteristics as [`std::collections::BTreeMap`] but
+//! with slightly different trade-offs as it's implemented on top of a flat storage (e.g. Vec).
+//!
+//! * [`LiteMap`] iteration is generally faster than `BTreeMap` because of the flat storage.
+//! * [`LiteMap`] can be pre-allocated whereas `BTreeMap` can't.
+//! * [`LiteMap`] has a smaller memory footprint than `BTreeMap` for small collections (< 20 items).
+//! * Lookup is `O(log(n))` like `BTreeMap`.
+//! * Insertion is generally `O(n)`, but optimized to `O(1)` if the new item sorts greater than the current items. In `BTreeMap` it's `O(log(n))`.
+//! * Deletion is `O(n)` whereas `BTreeMap` is `O(log(n))`.
+//! * Bulk operations like `from_iter`, `extend` and deserialization are optimized for inputs that
+//!    are already ordered but have `O(n*log(n))` complexity regardless of the order of the input items.
 //!
 //! ## Pluggable Backends
 //!