--- a/utils/litemap/src/lib.rs
+++ b/utils/litemap/src/lib.rs
@@ -26,8 +26,8 @@
 //! * Lookup is `O(log(n))` like `BTreeMap`.
 //! * Insertion is generally `O(n)`, but optimized to `O(1)` if the new item sorts greater than the current items. In `BTreeMap` it's `O(log(n))`.
 //! * Deletion is `O(n)` whereas `BTreeMap` is `O(log(n))`.
-//! * Bulk operations like `from_iter`, `extend` and deserialization are optimized for inputs that
-//!    are already ordered but have `O(n*log(n))` complexity regardless of the order of the input items.
+//! * Bulk operations like `from_iter`, `extend` and deserialization have an optimized `O(n)` path
+//!    for inputs that are ordered and `O(n*log(n))` complexity otherwise.
 //!
 //! ## Pluggable Backends
 //!