@arthurprs 2025-03-16
All concerns have been addressed.

Since we're doing breaking changes, I also included 
fb22056 to remove the `S: Serialize` bound but because of the higher-ranked-trait bounds, we hit a [borrow checker limitation in zerotrie](https://github.com/unicode-org/icu4x/actions/runs/13883257767/job/38844595550?pr=6132), which required 54c22cd. I suppose I could put back a manual loop if that's undesirable.

I'd appreciate if we could get this merged and a 0.8 release (not included)