@arthurprs 2025-04-07
Gentle ping. It would be good to get this PR wrapped up and the crate version bumped to `0.8`